% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n2o_flux.R
\name{calculate_n2o_flux}
\alias{calculate_n2o_flux}
\title{Calculate N2O fluxes from processed json files}
\usage{
calculate_n2o_flux(
  data,
  deadband = 30,
  deadband_c = 0,
  stop_time_ag = 120,
  offset_k = "json",
  opt_db = "no",
  use_json_parameters = 0
)
}
\arguments{
\item{use_json_parameters}{}
}
\description{
This function calculate N2O fluxes using the table created by process_json_files() using a linear and nonlinear regression.
It also allows to optimize the deadband within a given range and select the best fit between the two regressions.
}
\examples{
calculate_n2o_flux()
}
