% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Show results of linear and non-linear fits for N2O data, along with deadband and outliers when used}
\usage{
plot_fit(
  id,
  data,
  deadband = 30,
  stop_time_ag = 120,
  offset_k = "json",
  opt_db = "no",
  clean = "no"
)
}
\arguments{
\item{id}{The ID (written as date.LABEL) of measurements to plot. Several measurements can be processed at the same time.}

\item{data}{The data containing the measurements seconds by seconds, given by the process_json_files function.}

\item{deadband}{The desired value (sec) for the deadband. N2O observations before that time will be discarded.}

\item{stop_time_ag}{The duration of observation (sec).}

\item{offset_k}{The value of the offset (cm) for the collar. If set to "json", the offset will be retrieved from the json file.}

\item{opt_db}{The range at which the deadband should be optimized. For intance, if opt_db is set to '20-50', the optimum deadband will tested for all values between 20 and 50sec. The deadband won't be optimzed if set to 'no' (default).}

\item{clean}{Value (clean) used in the following formula to detect outliers and discard them. Outliers are defined as observations (obs_i) for which: obs_i > upper_quantile_0.75 + Inter_Quantile x (clean) OR obs_i < lower_quantile_0.25 - Inter_Quantile x (clean). Those observations are discarded. Outliers are not removed by default.}
}
\value{
A list containing the plots for all IDs
}
\description{
Show results of linear and non-linear fits for N2O data, along with deadband and outliers when used
}
\examples{
plot_res <- plot_n2o_flux<-calculate_n2o_flux(id="20250310.plot10",data=n2o_obs,opt_db="20-50",clean=2)
}
