% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_json_files.R
\name{process_json_files}
\alias{process_json_files}
\title{Processing json files from the Licor 8200 Smart Chamber}
\usage{
process_json_files(filepath)
}
\arguments{
\item{filepath}{Path to the json file. This argument can also contain multiple paths.}
}
\value{
A table containing all the observations and informations from the json file.
}
\description{
This function process the json file from the Licor 8200 Smart Chamber and return the result as a table that can be easily read and processed by R.
}
\examples{
n2o_obs<-process_json_files("filepath")
}
