% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_json_files.R
\name{process_json_files}
\alias{process_json_files}
\title{Processing json files from the Licor 8200 Smart Chamber}
\usage{
process_json_files(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Processing json files from the Licor 8200 Smart Chamber
}
\examples{
process_json_files("filepath")
}
